Factor out file for bigfiles
