+++ metrics/mypy_high_water_mark
@@ -1,1 +1,1 @@
-92.4100
+92.4300
